Implementation of the travelling distance constraint

+++ src/main/java/fr/cirad/domain/DistanceMatrix.java
@@ -4,18 +4,18 @@ import java.util.List;
 
 public class DistanceMatrix {
     public List<String> locations;
-    public Double[][] distances;
+    public Integer[][] distances;
 
     public DistanceMatrix() {
         // must have a no-args constructor so it can be deserialized by Jackson
     }
 
-    public DistanceMatrix(List<String> locations, Double[][] distances) {
+    public DistanceMatrix(List<String> locations, Integer[][] distances) {
         this.locations = locations;
         this.distances = distances;
     }
 
-    public Double getDistance(String location1, String location2) {
+    public Integer getDistance(String location1, String location2) {
         int index1 = locations.indexOf(location1);
         int index2 = locations.indexOf(location2);
         return distances[index1][index2];
